# api/openapi.yaml
openapi: 3.1.0
info:
  title: Mission Control API
  description: |
    Mission Control REST API for managing playbooks, notifications,
    configurations, views, and system resources.
  version: 1.0.0
  contact:
    name: Flanksource
    url: https://flanksource.com
servers:
  - url: /
    description: Current server
...
security:
  - BearerAuth: []
paths:
...
  /notification/summary:
    post:
      tags: [Notifications]
      summary: Get notification history summary
      description: Get summary of notification send history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSendHistorySummaryRequest"
      responses:
        "200":
          description: Notification summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSendHistorySummaryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /notification/silence:
    post:
      tags: [Notifications]
      summary: Create notification silence
      description: Create a silence rule for notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SilenceSaveRequest"
      responses:
        "200":
          description: Silence created
        "400":
          $ref: "#/components/responses/BadRequest"
  /notification/silence_preview:
    get:
      tags: [Notifications]
      summary: Preview notification silence
      description: Preview what notifications would be silenced
      parameters:
        - name: id
          in: query
          schema:
            type: string
            format: uuid
          description: Resource ID
        - name: type
          in: query
          schema:
            type: string
            enum: [config, component, check]
          description: Resource type
        - name: recursive
          in: query
          schema:
            type: boolean
          description: Include child resources
        - name: filter
          in: query
          schema:
            type: string
          description: CEL filter expression
        - name: selectors
          in: query
          schema:
            type: string
          description: JSON-encoded resource selectors
      responses:
        "200":
          description: Preview of silenced notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationSilencePreviewItem"
        "400":
          $ref: "#/components/responses/BadRequest"

  # View endpoints
...
  /application/{namespace}/{name}:
    get:
      tags: [Applications]
      summary: Get application
      description: Get application details by namespace and name
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          $ref: "#/components/responses/NotFound"
  /application/{namespace}/{name}/export:
    get:
      tags: [Applications]
      summary: Export application
      description: Export application in the requested format
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          required: false
          description: "Output format (default: json)"
          schema:
            type: string
            enum: [json, html, pdf, facet-html, facet-pdf]
            default: json
        - name: disposition
          in: query
          required: false
          description: "When set to 'attachment', includes a Content-Disposition header for file download"
          schema:
            type: string
            enum: [inline, attachment]
      responses:
        "200":
          description: Exported application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
            text/html:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"

  # Redirect endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication (for webhooks)
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
...
  schemas:
...
    HTTPError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        data: {}
...
    NotificationSendHistorySummaryRequest:
      type: object
      properties:
        groupBy:
          type: array
          items:
            type: string
        status:
          type: object
          properties:
            match:
              type: string
        resourceType:
          type: string
        tags:
          type: string
        search:
          type: string
        from:
          type: string
        to:
          type: string
        includeDeletedResources:
          type: boolean
        pageIndex:
          type: integer
        pageSize:
          type: integer
    NotificationSendHistorySummaryResponse:
      type: object
      properties:
        total:
          type: integer
        results: {}
    SilenceSaveRequest:
      type: object
      properties:
        name:
          type: string
        config_id:
          type: string
          format: uuid
        canary_id:
          type: string
          format: uuid
        check_id:
          type: string
          format: uuid
        component_id:
          type: string
          format: uuid
        from:
          type: string
        until:
          type: string
        description:
          type: string
        recursive:
          type: boolean
        filter:
          type: string
        selectors:
          type: array
          items:
            $ref: "#/components/schemas/ResourceSelector"
    NotificationSilencePreviewItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        notification_id:
          type: string
          format: uuid
        source_event:
          type: string
        resource_id:
          type: string
          format: uuid
        status:
          type: string
        resource:
          type: object
...
    ResourceSelector:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        types:
          type: array
          items:
            type: string
        statuses:
          type: array
          items:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        field_selector:
          type: string
...
    Application:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        namespace:
          type: string
        description:
          type: string
        properties:
          type: array
          items:
            type: object
        access_control:
          type: object
        incidents:
          type: array
          items:
            type: object
        locations:
          type: array
          items:
            type: object
        backups:
          type: array
          items:
            type: object
        restores:
          type: array
          items:
            type: object
        findings:
          type: array
          items:
            type: object
        sections:
          type: array
          items:
            type: object
...